initial commit
This is a standalone version of systemd-bootchart, extracted from the
systemd repository as of commit ID 2c45295e4764.

The idea is to give this tool a new home, because it is not actually related
to systemd in any way, can well live outside of the repository, and is
actually useful beyond the scope of systemd itself.