work correctly without CONFIG_SCHED_DEBUG
The app appeared to work without CONFIG_SCHED_DEBUG (no error messages),
but the PS part of the graph was not plotted, and process command lines
were not collected with -C. It was hard to find out what's the reason. A
small change fixes the code to work correctly without requirement for
this config.