Don't take time from /proc/schedstat for processes.
I assume this is a copy-paste error. We shouldn't take the rt and wt
values here from /proc/schedstat, as they're resulting in entirely
incorrect values. Instead, the loop code fills them properly with
the per-process value already, and they can be left blank here
intentionally. This was the case before in the pre-systemd code tree.