Scale bi/bo graphs independently.
The bi/bo graphs have, since the beginning of this project, always
been scaled with the same factor. This meant that 'bo' graphs were
generally always entirely empty, and thefore very hard to spot any
issues with.

In general, especially during boot, we don't want to see any disk
io being written to disk, since it's much slower than reading no
matter the storage.

For these reasons, it doesn't make any sense to scale the 'bo' graph
with the same scale as 'bi'. Instead, we scale bi/bo graphs from now
on independently, and this will make it a lot easier to spot issues
with writes and clearly identify them.

Because the scale is now only dependent on 'bo' data, you should
see the peak 'bo' value as well appear with a label in the graph
showing you the peak value. This was previously hidden due to the
fact that most of the time 'bo' was such a small value compared
to 'bi'.