Keep working with PID#1
Do not skip the first sample for PID#1

The continue statement was introduced very early with commit 5da71e1
in the original reposiotory before bootchart was merged into systemd,
when there was little code at the end of the while loop and apparently
skipping it didn't hurt. Specifically commit 9b004a7 introduces
still_runnig flag which is never set for PID#1 which makes it subject
to garbage collection in garbage_collect_dead_processes().