Update configuration/options summary in svg
Some of the comments summarizing the configuration/options could be
confusing.  E.g., "f" was used for "Filtering", but "-f" is the
command-line option used to specify the sample frequency.  Also, "hz"
was used but there is no command-line option or config directive that
mentions "hertz".

Change the summmary to use the names from the config file.  Update the
sample config file so it is more obvious what is parsed as a float,
and also include "Cmdline", which was missing.